fixed registration

+++ src/php/functions/deleteOrder.php
@@ -9,9 +9,11 @@ if(isset($_SESSION['user']) && isset($_GET['id']) && isset($_GET['session']) && 
         return;
     }
 
-    $query = oci_parse($connection,"DELETE FROM PURCHASES WHERE USER_EMAIL=:email AND ID=:id");
+    $query = oci_parse($connection,"DELETE FROM PURCHASES WHERE USER_EMAIL=:email AND ID=:id AND PURCHASE_STATE!=:forbidden_state");
     oci_bind_by_name($query,":email",$_GET['session']);
     oci_bind_by_name($query,":id",$_GET['id']);
+    $forbiddenState = 3;
+    oci_bind_by_name($query,":forbidden_state",$forbiddenState);
     if(oci_execute($query)){
         oci_commit($connection);
         header('Location: /php/userinfo.php');
